MCS-68 MCS-432 Split the integration tests into three runnable files. Stubbed out code for the future prefab tests. Updated the README for integration tests.
