+++ integration_tests/README.md
@@ -1,31 +1,43 @@
 # MCS Integration Tests
 
-The goal of the integration tests is to verify that specific sequences of actions entered via the MCS Python API lead to specific metadata outputs containing expected observation, navigation, and object interaction results.
+The goal of the integration tests is to:
+- Verify specific sequences of actions entered via the MCS Python API return specific metadata outputs containing expected observation, navigation, and object interaction results within tightly-controlled scenes (the "Handmade" tests).
+- Verify all objects that should be compiled into the given MCS Unity build actually exist in that build (the "Prefab" tests). Because the files for some objects (Unity Prefabs) are stored in a private repository and must be copied into the Resources folder before building, we want to verify that each object was correctly included.
 
-## Run All Test Scenes
+## Run Handmade Tests
+
+```
+python run_handmade_tests.py <mucs_unity_build_file_path>
+```
+
+### Run Specific Metadata Tier
+
+```
+python run_handmade_tests.py <mucs_unity_build_file_path> --metadata <level1|level2|oracle>
+```
+
+### Run Specific Test Scene
+
+```
+python run_handmade_tests.py <mucs_unity_build_file_path> --test 001
+```
+
+Replace `001` with any test scene.
+
+## Run Prefab Tests
+
+TODO MCS-432
+
+## Run All Tests
 
 ```
 python run_tests.py <mucs_unity_build_file_path>
 ```
 
-## Run Specific Metadata Tier
-
-```
-python run_tests.py <mucs_unity_build_file_path> --metadata <level1|level2|oracle>
-```
-
-## Run Specific Test Scene
-
-```
-python run_tests.py <mucs_unity_build_file_path> --test 001
-```
-
-Replace `001` with any test scene.
-
-## Adding New Test Scenes
+## Adding New Handmade Test Scenes
 
 Create the following files (Replace `<X>` with your test name):
 
-- `./data/<X>.scene.json`: MCS scene JSON file
-- `./data/<X>.actions.txt`: List of actions to take in the scene, with the following format: `action<,parameter=value>\n`
-- `./data/<X>.level1.outputs.json`, `./data/<X>.level2.outputs.json`, `./data/<X>.oracle.outputs.json`: JSON list of expected step and object output metadata per action step, starting at step 0 (initialization). Please do not skip any step. See the Python code for the available JSON test properties. One separate file per metadata tier.
+- `./data/<X>.scene.json`: An MCS scene configuration JSON file. Give it a useful `name` that describes the specific test case for future developers.
+- `./data/<X>.actions.txt`: The list of specific actions to take in the scene, each action and its corresponding parameters on its own line, with the following format: `action<,parameter=value>\n`
+- `./data/<X>.<level1|level2|oracle>.outputs.json`: Each file is a JSON list of expected step and object output metadata per action step, starting at step 0 (initialization), for runs with the file's corresponding metadata tier. Please do not skip any step. See the Python code for the available JSON test properties. One separate file per metadata tier, and each metadata tier should have its own corresponding file.
